// frontend/src/app/models/relation.ts
export interface Relation {
    contractor_id: string
    employee_id: string
    contractor: string
    employee: string
    identification: number
    contract_type: string
    worked_time: number
    working_time: number
    start: string
    work_name: string
    createdAt?: string
    updatedAt?: string
    _id?: string
}